Fix pay-in Sankey amount formatting
